+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-07-08
+### Added
+- Table columns can be sorted.
+- Matched search text is highlighted.
+- More details shown on sheet modal.
+
+### Changed
+- All references to "POC" changed to "Author".
+- Filter boxes moved to column header.
+
 ## [0.1.0] - 2020-06-10
 ### Added
 - Initial release.